% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{token}
\alias{token}
\title{read my encrypted token}
\usage{
token(key, pkey = NULL)
}
\arguments{
\item{key}{For symmetric decryption, 'pkey' should be NULL (default) and
'key' can be either a string (Default is 'pass') or a raw object. For
asymmetric decryption, both 'key' (private key of the decrypter) and 'pkey'
(public key of the encrypter) should be raw objects.}

\item{pkey}{See 'key'}
}
\value{
decrypted token.
}
\description{
read my encrypted token
}
